mysite/requirements.txt
boto==2.42.0
configparser==3.5.0
coreapi==2.0.0
Django==1.10.1
django-autofixture==0.12.1
django-bootstrap3==7.0.1
django-filter==0.14.0
django-formtools==1.0
django-rest-swagger==2.0.5
django-storages==1.5.1
djangorestframework==3.4.6
enum34==1.1.6
flake8==3.0.4
invoke==0.13.0
itypes==1.1.0
Markdown==2.6.6
mccabe==0.5.2
openapi-codec==1.0.0
Pillow==3.3.1
pycodestyle==2.0.0
pyflakes==1.2.3
requests==2.11.1
simplejson==3.8.2
uritemplate==3.0.0
